+++ CHANGELOG.md
@@ -45,6 +45,7 @@ Changelog entries are classified using the following labels _(from [keep-a-chang
 
 ### Added
 
+- Adds support for `options.history` on the `Input` and `Text` prompts. 
 - Adds support for `options.term` to set the terminal, thanks to [@tunnckoCore](https://github.com/tunnckoCore). At the moment this is only used in a couple of edge cases with the `Survey` and `MultiScale` prompts to check if the terminal is Hyper.
 - `options.skip` may now be a Boolean, thanks to [@tunnckoCore](https://github.com/tunnckoCore)